% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{getSimpsonBetaConditional}
\alias{getSimpsonBetaConditional}
\title{Title}
\usage{
getSimpsonBetaConditional(
  Y,
  tab,
  n.s.tab,
  pop1,
  pop2,
  c.top,
  d.top,
  c.loc1,
  d.loc1,
  c.loc2,
  d.loc2
)
}
\arguments{
\item{Y}{Species table: rows are local populations, columns are species}

\item{d.loc2}{}
}
\value{

}
\description{
Title
}
