% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{getSimpsonBetaUnconditional}
\alias{getSimpsonBetaUnconditional}
\title{Calculate unconditional Simpson's index for beta diversity}
\usage{
getSimpsonBetaUnconditional(c.top, d.top)
}
\arguments{
\item{c.top}{Top-level concentration parameter}

\item{d.top}{Top-level discount parameter}
}
\value{
Unconditional Simpson's index for beta diversity
}
\description{
Calculate unconditional Simpson's index for beta diversity
}
