% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{getSimpsonConditional}
\alias{getSimpsonConditional}
\title{Title}
\usage{
getSimpsonConditional(Y, tab, n.s.tab, pop, c.top, d.top, c.loc, d.loc)
}
\arguments{
\item{d.loc}{}
}
\value{

}
\description{
Title
}
