% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{getSimpsonUnconditional}
\alias{getSimpsonUnconditional}
\title{Calculate unconditional Simpson's index for a particular population}
\usage{
getSimpsonUnconditional(c.top, d.top, c.loc, d.loc)
}
\arguments{
\item{c.top}{Top-level concentration parameter}

\item{d.top}{Top-level discount parameter}

\item{c.loc}{Local-level concentration parameter for the population of interest}

\item{d.loc}{Local-level discount parameter for the population of interest}
}
\value{
Unconditional Simpson's index
}
\description{
Calculate unconditional Simpson's index for a particular population
}
