% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpy_functions.R
\name{hpySampler}
\alias{hpySampler}
\title{Hierarchical Pitman-Yor Gibbs sampler}
\usage{
hpySampler(
  Y,
  n.iter,
  n.burn,
  p.shape,
  p.scale,
  p.shape.pop = NULL,
  p.scale.pop = NULL,
  type = c("py", "dp"),
  quiet = FALSE
)
}
\arguments{
\item{Y}{Matrix of taxon counts (rows are populations, columns are species)}

\item{n.iter}{Number of MCMC iterations (after burn-in)}

\item{n.burn}{Number of burn-in iterations}

\item{p.shape}{Shape parameter for gamma prior on top-level concentration parameter}

\item{p.scale}{Scale parameter for gamma prior on top-level concentration parameter}

\item{p.shape.pop}{Vector with length equal to number of populations containing shape parameters for gamma priors on population-level concentration parameter}

\item{p.scale.pop}{Vector with length equal to number of populations containing scale parameters for gamma priors on population-level concentration parameter}

\item{type}{Type of hierarchical process: "py" for Pitman-Yor process, "dp" for Dirichlet process}

\item{quiet}{Set to TRUE to suppress console output}
}
\value{
Returns a list containing the following elements:
\itemize{
\item gamma, alpha, theta, sigma - Vectors containing posterior samples for each of the HPY parameters.
\item tab - List with table indicators for each population.
\item t.c - Table counts (and species corresponding to each table)
\item n.tab - Number of tables in each population
\item n.s.tab - Number of tables for a given species in each population
}
}
\description{
Hierarchical Pitman-Yor Gibbs sampler
}
